repost window wip
