+++ src/assets/styles/index.scss
@@ -8,7 +8,6 @@
 @import "form.scss";
 @import "heading.scss";
 @import "icon.scss";
-@import "list.scss";
 @import "table.scss";
 @import "text.scss";
 @import "tooltip.scss";
