--- src/assets/styles/list.scss
@@ -1,44 +1,0 @@
-.nn-list {
-  margin: 0;
-  padding: 0;
-
-  & > li {
-    margin: 0;
-    padding: 4px 8px;
-    border-top: 1px solid #ddd;
-    &:first-child {
-      border-top: none;
-    }
-  }
-}
-
-.nn-table-list {
-  display: table;
-  & > li {
-    display: table-row;
-  }
-}
-
-.nn-dlist {
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  margin: 0;
-  padding: 0;
-
-  & > dt {
-    margin: 0;
-    padding: 8px 0 0;
-    font-weight: bold;
-    line-height: 1.5em;
-    border-top: 1px solid #ddd;
-    &:first-child {
-      border-top: none;
-    }
-  }
-  & > dd {
-    margin: -1.5em 0 0 auto;
-    padding: 0 0 8px;
-    line-height: 1.5em;
-  }
-}
