+++ src/components/MastodonNotification.vue
@@ -124,7 +124,7 @@ const openUserPage = (user: MastodonToot["account"]) => {
   position: relative;
   width: 100%;
   margin: 0;
-  padding: 4px 8px;
+  padding: 8px;
   background-color: transparent;
 
   &.indent-1 {
