+++ src/router.ts
@@ -33,6 +33,11 @@ const routes = [
         name: "PostReaction",
         component: () => import("./pages/post/reaction.vue"),
       },
+      {
+        path: "/post/repost",
+        name: "PostRepost",
+        component: () => import("./pages/post/repost.vue"),
+      },
     ],
   },
   {
